+++ src/main/java/expresscarts/config/BlockTypeAdapter.java
@@ -1,0 +1,22 @@
+package expresscarts.config;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.block.Block;
+
+import java.io.IOException;
+
+public class BlockTypeAdapter extends TypeAdapter<Block> {
+    @Override
+    public void write(JsonWriter writer, Block block) throws IOException {
+        writer.value(BuiltInRegistries.BLOCK.getKey(block).toString());
+    }
+
+    @Override
+    public Block read(JsonReader reader) throws IOException {
+        return BuiltInRegistries.BLOCK.getOptional(ResourceLocation.parse(reader.nextString())).orElseThrow(IOException::new);
+    }
+}