+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Improve the display of the `/expresscarts` command
+- Improved the display of the `/expresscarts` command.
+- Added configurable max speed multipliers based on the block underneath the rails. Set these using `/config expresscarts blockSpeedMultipliers put <block> <multiplier>`.